% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_preface.R.R
\name{$.pack_of_constants}
\alias{$.pack_of_constants}
\title{Extract constant from pack}
\usage{
\method{$}{pack_of_constants}(pack, name)
}
\arguments{
\item{pack}{pack_of_constants}

\item{name}{target constant

This function differs from the base list extraction method in that it avoids partial matching of keys and throws
an error if the looked-for constant is not contained within the pack.}
}
\description{
Extract constant from pack
}
\keyword{internal}
