% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_spiderplot.R
\name{mock_spiderplot_mm}
\alias{mock_spiderplot_mm}
\title{Mock Application for Spider Plot Module}
\usage{
mock_spiderplot_mm()
}
\value{
Launches a Shiny application with the spider plot module
}
\description{
Creates a mock application demonstrating the spider plot module functionality.
}
