% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spiderplot.R
\name{spiderplot_UI}
\alias{spiderplot_UI}
\title{Spider Plot Module UI}
\usage{
spiderplot_UI(id, show_color_vars, show_facet_rows, show_facet_cols)
}
\arguments{
\item{id}{`[character(1)]`. 
Unique identifier for the module UI.}

\item{show_color_vars}{`[logical(1)]`. 
Whether to show the color variable selection.}

\item{show_facet_rows}{`[logical(1)]`. 
Whether to show the facet rows selection.}

\item{show_facet_cols}{`[logical(1)]`. 
Whether to show the facet columns selection.}
}
\description{
Spider Plot Module UI
}
